making exit lockup time as module parameter

+++ x/amm/migrations/v11_migration.go
@@ -6,7 +6,7 @@ import (
 
 func (m Migrator) V11Migration(ctx sdk.Context) error {
 	legacyParams := m.keeper.GetLegacyParams(ctx)
-	legacyParams.LpLockupDuration = uint64(3600)
+	legacyParams.LpLockupDuration = 3600
 
 	m.keeper.SetParams(ctx, legacyParams)
 	return nil
